simplified plugin template
