+++ .gitignore
@@ -1,0 +1,7 @@
+*.suo
+*.sdf
+*.opensdf
+bin/
+Win32/
+x64/
+release/
