+++ pluginmain.cpp
@@ -1,9 +1,9 @@
 //modded plugin script
-// original made by Tomer Zait
+//original made by Tomer Zait
 //modded by storm shadow
 
 #include "pluginmain.h"
-#include "py.h"
+#include "plugin.h"
 
 int pluginHandle;
 HWND hwndDlg;
@@ -11,6 +11,7 @@ int hMenu;
 int hMenuDisasm;
 int hMenuDump;
 int hMenuStack;
+HINSTANCE hInst;
 
 DLL_EXPORT bool pluginit(PLUG_INITSTRUCT* initStruct)
 {
@@ -18,7 +19,6 @@ DLL_EXPORT bool pluginit(PLUG_INITSTRUCT* initStruct)
     initStruct->sdkVersion = PLUG_SDKVERSION;
     strcpy(initStruct->pluginName, plugin_name);
     pluginHandle = initStruct->pluginHandle;
-    pyInit(initStruct);
     return true;
 }
 
@@ -41,8 +41,6 @@ DLL_EXPORT void plugsetup(PLUG_SETUPSTRUCT* setupStruct)
 extern "C" DLL_EXPORT BOOL APIENTRY DllMain(HINSTANCE hinstDLL, DWORD fdwReason, LPVOID lpvReserved)
 {
     if(fdwReason == DLL_PROCESS_ATTACH)
-    {
         hInst = hinstDLL;
-    }
     return TRUE;
 }
