+++ x64dbgpy-template.vcxproj
@@ -12,8 +12,7 @@
   </ItemGroup>
   <ItemGroup>
     <ClCompile Include="pluginmain.cpp" />
-    <ClCompile Include="py.cpp" />
-    <ClCompile Include="stringutils.cpp" />
+    <ClCompile Include="plugin.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="pluginmain.h" />
@@ -78,9 +77,8 @@
     <ClInclude Include="pluginsdk\_scriptapi_module.h" />
     <ClInclude Include="pluginsdk\_scriptapi_pattern.h" />
     <ClInclude Include="pluginsdk\_scriptapi_register.h" />
-    <ClInclude Include="py.h" />
+    <ClInclude Include="plugin.h" />
     <ClInclude Include="resource.h" />
-    <ClInclude Include="stringutils.h" />
   </ItemGroup>
   <ItemGroup>
     <Reference Include="System" />
@@ -90,14 +88,14 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <ResourceCompile Include="x64dbg-python-template.rc" />
+    <ResourceCompile Include="x64dbgpy-template.rc" />
   </ItemGroup>
   <ItemGroup>
     <None Include="res\python.png" />
   </ItemGroup>
   <PropertyGroup Label="Globals">
     <Keyword>Win32Proj</Keyword>
-    <ProjectName>x64dbg-python-plugin-template</ProjectName>
+    <ProjectName>x64dbgpy-template</ProjectName>
     <ProjectGuid>{E7BA66E1-677E-D937-AC50-73AB6ADA89E0}</ProjectGuid>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.Default.props" />
@@ -105,12 +103,12 @@
     <ConfigurationType>DynamicLibrary</ConfigurationType>
     <UseDebugLibraries>false</UseDebugLibraries>
     <CharacterSet>Unicode</CharacterSet>
-    <PlatformToolset>v120</PlatformToolset>
+    <PlatformToolset>v120_xp</PlatformToolset>
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'" Label="Configuration">
     <ConfigurationType>DynamicLibrary</ConfigurationType>
     <UseDebugLibraries>false</UseDebugLibraries>
-    <PlatformToolset>v120</PlatformToolset>
+    <PlatformToolset>v120_xp</PlatformToolset>
   </PropertyGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.props" />
   <ImportGroup Label="ExtensionSettings">
@@ -124,9 +122,9 @@
   <PropertyGroup Label="UserMacros" />
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|Win32'">
     <LinkIncremental>false</LinkIncremental>
-    <OutDir>$(SolutionDir)plugins\</OutDir>
+    <OutDir>$(SolutionDir)bin\x32\</OutDir>
     <IntDir>$(Platform)\$(Configuration)\</IntDir>
-    <TargetName>x64dbg_Editor</TargetName>
+    <TargetName>$(ProjectName)</TargetName>
     <TargetExt>.dp32</TargetExt>
     <IncludePath>$(Python27x86)\include;$(IncludePath)</IncludePath>
     <LibraryPath>$(Python27x86)\libs;$(LibraryPath)</LibraryPath>
@@ -136,11 +134,11 @@
   </PropertyGroup>
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Release|x64'">
     <LinkIncremental>false</LinkIncremental>
-    <OutDir>$(SolutionDir)plugins\</OutDir>
-    <TargetName>x64dbg-editor</TargetName>
+    <OutDir>$(SolutionDir)bin\x64\</OutDir>
+    <TargetName>$(ProjectName)</TargetName>
     <TargetExt>.dp64</TargetExt>
-    <IncludePath>$(Python27x64)\include;$(IncludePath)</IncludePath>
-    <LibraryPath>$(Python27x64)\libs;$(LibraryPath)</LibraryPath>
+    <IncludePath>$(IncludePath)</IncludePath>
+    <LibraryPath>$(LibraryPath)</LibraryPath>
     <PostBuildEventUseInBuild>false</PostBuildEventUseInBuild>
     <PreLinkEventUseInBuild>false</PreLinkEventUseInBuild>
     <PreBuildEventUseInBuild>false</PreBuildEventUseInBuild>
@@ -159,7 +157,7 @@
       <SubSystem>Windows</SubSystem>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>psapi.lib;pluginsdk\x32dbg.lib;pluginsdk\x32bridge.lib;pluginsdk\TitanEngine\TitanEngine_x86.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>psapi.lib;pluginsdk\x32dbg.lib;pluginsdk\x32bridge.lib;%(AdditionalDependencies)</AdditionalDependencies>
       <AdditionalLibraryDirectories>C:\Python27\Libs</AdditionalLibraryDirectories>
     </Link>
     <PostBuildEvent>
@@ -192,7 +190,7 @@
       <SubSystem>Windows</SubSystem>
       <EnableCOMDATFolding>true</EnableCOMDATFolding>
       <OptimizeReferences>true</OptimizeReferences>
-      <AdditionalDependencies>psapi.lib;pluginsdk\x64dbg.lib;pluginsdk\x64bridge.lib;pluginsdk\TitanEngine\TitanEngine_x64.lib;C:\x64dbg\x64\plugins\pyx64\libs\python27.lib;%(AdditionalDependencies)</AdditionalDependencies>
+      <AdditionalDependencies>psapi.lib;pluginsdk\x64dbg.lib;pluginsdk\x64bridge.lib;%(AdditionalDependencies)</AdditionalDependencies>
     </Link>
     <PostBuildEvent>
       <Command>
