+++ x64dbgpy_plugins/x64dbgpy-template.py
@@ -1,0 +1,1 @@
+print 'This is just a template, do cool stuff here...'
